generate_code.py
import ollama
import os
import yaml

def load_config(config_path):
    """Load training configuration from YAML file"""
    with open(config_path, 'r') as file:
        return yaml.safe_load(file)

def generate_code_with_qwen(prompt, model="qwen2.5-coder:1.5b"):
    """Generate code using Qwen2.5-Coder via Ollama"""
    print(f"Generating code with {model}...")
    response = ollama.generate(
        model=model,
        prompt=prompt
    )
    return response['response']

def main():
    # Load configuration
    config = load_config("configs/training_config.yaml")
    
    # Create directories if they don't exist
    os.makedirs("src/models", exist_ok=True)
    os.makedirs("src/data", exist_ok=True)
    os.makedirs("src/training", exist_ok=True)
    os.makedirs("src/utils", exist_ok=True)
    os.makedirs("notebooks", exist_ok=True)
    os.makedirs("tests", exist_ok=True)
    
    # Generate PolymorphicSentimentAnalyzer model
    model_prompt = f"""
    Write a Python class named PolymorphicSentimentAnalyzer that inherits from torch.nn.Module.
    This class should implement an advanced sentiment analyzer incorporating polymorphic research insights.
    
    Requirements:
    1. Use QwenForSequenceClassification as the backbone model
    2. Implement polymorphic-inspired feature layers
    3. Include a multi-stage classification head
    4. Add methods for loading pretrained models and making predictions
    5. Include proper error handling and documentation
    
    Configuration details:
    - Model name: {config['model']['name']}
    - Number of labels: {config['model']['num_labels']}
    - Max length: {config['model']['max_length']}
    
    The class should have the following structure:
    - __init__ method to initialize the model
    - forward method for forward pass
    - predict method for making predictions
    - load_pretrained method for loading saved models
    - get_confidence method for getting prediction confidence
    """
    
    model_code = generate_code_with_qwen(model_prompt)
    
    with open("src/models/qwen_sentiment_model.py", "w") as f:
        f.write(model_code)
    
    print("PolymorphicSentimentAnalyzer model saved to src/models/qwen_sentiment_model.py")
    
    # Generate polymorphic features module
    features_prompt = """
    Write a Python module for polymorphic feature extraction.
    Create a class named PolymorphicFeatureExtractor that implements feature engineering approaches 
    from recent polymorphic malware detection research.
    
    Requirements:
    1. Implement structural feature engineering techniques
    2. Include dynamic feature extraction methods
    3. Add behavioral analysis capabilities
    4. Include methods for feature preprocessing and normalization
    5. Add proper documentation and error handling
    
    The class should have methods for:
    - extract_structural_features
    - extract_dynamic_features
    - extract_behavioral_features
    - combine_features
    - normalize_features
    """
    
    features_code = generate_code_with_qwen(features_prompt)
    
    with open("src/models/polymorphic_features.py", "w") as f:
        f.write(features_code)
    
    print("PolymorphicFeatureExtractor module saved to src/models/polymorphic_features.py")
    
    # Generate data preprocessing module
    preprocessing_prompt = """
    Write a Python module for data preprocessing in sentiment analysis.
    Create functions for preprocessing text data with polymorphic-inspired techniques.
    
    Requirements:
    1. Implement text cleaning and normalization
    2. Include polymorphic preprocessing pipeline
    3. Add functions for loading and processing datasets
    4. Include feature engineering functions
    5. Add proper documentation and error handling
    
    Include the following functions:
    - clean_text: for basic text cleaning
    - tokenize_text: for tokenizing text using BERT tokenizer
    - polymorphic_preprocessing_pipeline: advanced preprocessing incorporating polymorphic analysis techniques
    - load_dataset: for loading sentiment analysis datasets
    - create_data_loader: for creating PyTorch data loaders
    """
    
    preprocessing_code = generate_code_with_qwen(preprocessing_prompt)
    
    with open("src/data/preprocessing.py", "w") as f:
        f.write(preprocessing_code)
    
    print("Data preprocessing module saved to src/data/preprocessing.py")
    
    # Generate trainer module
    trainer_prompt = """
    Write a Python class named PolymorphicSentimentTrainer for training the polymorphic sentiment analysis model.
    
    Requirements:
    1. Implement training loop with configurable parameters
    2. Include evaluation and validation functionality
    3. Add model saving and loading capabilities
    4. Implement early stopping and learning rate scheduling
    5. Include comprehensive logging and progress tracking
    6. Add proper documentation and error handling
    
    The class should have methods for:
    - __init__: Initialize trainer with configuration
    - train: Main training loop
    - evaluate: Evaluate model on validation set
    - save_model: Save trained model
    - load_model: Load pretrained model
    - plot_training_history: Visualize training progress
    """
    
    trainer_code = generate_code_with_qwen(trainer_prompt)
    
    with open("src/training/trainer.py", "w") as f:
        f.write(trainer_code)
    
    print("PolymorphicSentimentTrainer module saved to src/training/trainer.py")
    
    print("Code generation complete!")

if __name__ == "__main__":
    main()