requirements.txt
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
flask>=2.0.0
pyyaml>=6.0
pytest>=6.2.0
jupyter>=1.0.0
notebook>=6.4.0